+++ cognite/neat/_graph/extractors/_dms.py
@@ -1,3 +1,4 @@
+import json
 import urllib.parse
 from collections.abc import Iterable, Iterator, Set
 from functools import cached_property
@@ -189,38 +190,47 @@ class DMSExtractor(BaseExtractor):
         for view_id, properties in instance.properties.items():
             namespace = self._get_namespace(view_id.space)
             for key, value in properties.items():
-                for predicate_str, object_ in self._get_predicate_objects_pair(key, value):
+                for predicate_str, object_ in self._get_predicate_objects_pair(key, value, self.unpack_json):
                     yield id_, namespace[urllib.parse.quote(predicate_str)], object_
 
-    def _get_predicate_objects_pair(self, key: str, value: PropertyValue) -> Iterable[tuple[str, Literal | URIRef]]:
+    def _get_predicate_objects_pair(
+        self, key: str, value: PropertyValue, unpack_json: bool
+    ) -> Iterable[tuple[str, Literal | URIRef]]:
         if isinstance(value, str | float | bool | int):
             yield key, Literal(value)
         elif isinstance(value, dict) and "space" in value and "externalId" in value:
             yield key, self._as_uri_ref(dm.DirectRelationReference.load(value))
-        elif isinstance(value, dict) and self.unpack_json:
-            for sub_key, sub_value in value.items():
-                if isinstance(sub_value, str):
-                    if sub_value.casefold() in self.empty_values:
-                        continue
-                    if self.str_to_ideal_type:
-                        yield sub_key, Literal(string_to_ideal_type(sub_value))
-                    else:
-                        yield sub_key, Literal(sub_value)
-                elif isinstance(sub_value, int | float | bool):
-                    yield sub_key, Literal(sub_value)
-                elif isinstance(sub_value, dict):
-                    yield from self._get_predicate_objects_pair(f"{key}_{sub_key}", sub_value)
-                elif isinstance(sub_value, list):
-                    for item in sub_value:
-                        yield from self._get_predicate_objects_pair(f"{key}_{sub_key}", item)
-                else:
-                    yield sub_key, Literal(str(sub_value))
+        elif isinstance(value, dict) and unpack_json:
+            yield from self._unpack_json(value)
         elif isinstance(value, dict):
             # This object is a json object.
-            yield key, Literal(str(value), datatype=XSD._NS["json"])
+            yield key, Literal(json.dumps(value), datatype=XSD._NS["json"])
         elif isinstance(value, list):
             for item in value:
-                yield from self._get_predicate_objects_pair(key, item)
+                yield from self._get_predicate_objects_pair(key, item, False)
+
+    def _unpack_json(self, value: dict, parent: str | None = None) -> Iterable[tuple[str, Literal | URIRef]]:
+        for sub_key, sub_value in value.items():
+            key = f"{parent}_{sub_key}" if parent else sub_key
+            if isinstance(sub_value, str):
+                if sub_value.casefold() in self.empty_values:
+                    continue
+                if self.str_to_ideal_type:
+                    yield key, Literal(string_to_ideal_type(sub_value))
+                else:
+                    yield key, Literal(sub_value)
+            elif isinstance(sub_value, int | float | bool):
+                yield key, Literal(sub_value)
+            elif isinstance(sub_value, dict):
+                yield from self._unpack_json(sub_value, key)
+            elif isinstance(sub_value, list):
+                for no, item in enumerate(sub_value, 1):
+                    if isinstance(item, dict):
+                        yield from self._unpack_json(item, f"{key}_{no}")
+                    else:
+                        yield from self._get_predicate_objects_pair(key, item, self.unpack_json)
+            else:
+                yield key, Literal(str(sub_value))
 
     def _as_uri_ref(self, instance: Instance | dm.DirectRelationReference) -> URIRef:
         return self._get_namespace(instance.space)[urllib.parse.quote(instance.external_id)]