+++ cognite/neat/_graph/queries/_base.py
@@ -1,8 +1,9 @@
+import urllib.parse
 from collections import defaultdict
 from collections.abc import Iterable
-from typing import Literal, cast, overload
+from typing import Any, Literal, cast, overload
 
-from rdflib import RDF, Dataset, Graph, Namespace, URIRef
+from rdflib import RDF, XSD, Dataset, Graph, Namespace, URIRef
 from rdflib import Literal as RdfLiteral
 from rdflib.graph import DATASET_DEFAULT_GRAPH_ID
 from rdflib.query import ResultRow
@@ -72,6 +73,20 @@ class Queries:
             cast(URIRef, type_): remove_namespace_from_uri(cast(URIRef, type_))
             for (type_,) in list(self.graph(named_graph).query(query))
         }
+
+    def properties_by_type(self, named_graph: URIRef | None = None) -> dict[URIRef, dict[URIRef, str]]:
+        """Properties and their short form in the graph by type
+
+        Args:
+            named_graph: Named graph to query over, default None (default graph)
+
+        """
+        query = """SELECT DISTINCT ?type ?property
+               WHERE {?s a ?type . ?s ?property ?o . FILTER(?property != rdf:type)}"""
+        properties_by_type: dict[URIRef, dict[URIRef, str]] = defaultdict(dict)
+        for type_, property_ in cast(ResultRow, list(self.graph(named_graph).query(query))):
+            properties_by_type[type_][property_] = remove_namespace_from_uri(property_)  # type: ignore[index]
+        return properties_by_type
 
     def property_uri(self, property_: str, named_graph: URIRef | None = None) -> list[URIRef]:
         """Get the URIRef of a property
@@ -168,7 +183,8 @@ class Queries:
         instance_type: URIRef | None = None,
         property_renaming_config: dict | None = None,
         named_graph: URIRef | None = None,
-    ) -> tuple[str, dict[str | InstanceType, list[str]]] | None:
+        remove_uri_namespace: bool = True,
+    ) -> tuple[URIRef, dict[str | InstanceType, list[Any]]] | None:
         """DESCRIBE instance for a given class from the graph store
 
         Args:
@@ -176,13 +192,13 @@ class Queries:
             instance_type: Type of the instance, default None (will be inferred from triples)
             property_renaming_config: Dictionary to rename properties, default None (no renaming)
             named_graph: Named graph to query over, default None (default graph)
+            remove_uri_namespace: Whether to remove the namespace from the URI, by default True
 
 
         Returns:
             Dictionary of instance properties
         """
-        property_values: dict[str, list[str]] = defaultdict(list)
-        identifier = remove_namespace_from_uri(instance_id, validation="prefix")
+        property_values: dict[str, list[str] | list[URIRef]] = defaultdict(list)
         for _, predicate, object_ in cast(list[ResultRow], self.graph(named_graph).query(f"DESCRIBE <{instance_id}>")):
             if object_.lower() in [
                 "",
@@ -205,29 +221,37 @@ class Queries:
                 property_ = RDF.type
                 renamed_property_ = property_
 
-            if isinstance(object_, URIRef):
+            value: Any
+            if isinstance(object_, URIRef) and remove_uri_namespace:
+                # These properties contain the space in the Namespace.
                 value = remove_namespace_from_uri(object_, validation="prefix")
+            elif isinstance(object_, URIRef):
+                value = object_
             elif isinstance(object_, RdfLiteral):
-                value = object_.toPython()
+                if object_.datatype == XSD._NS["json"]:
+                    # For JSON literals, the .toPython() returns a Literal object.
+                    value = str(object_)
+                else:
+                    value = object_.toPython()
             else:
                 # It is a blank node
                 value = str(object_)
 
             # add type to the dictionary
             if predicate != RDF.type:
-                property_values[renamed_property_].append(value)
+                property_values[renamed_property_].append(value)  # type: ignore[arg-type]
             else:
                 # guarding against multiple rdf:type values as this is not allowed in CDF
                 if RDF.type not in property_values:
                     property_values[RDF.type].append(
-                        remove_namespace_from_uri(instance_type, validation="prefix") if instance_type else value
+                        remove_namespace_from_uri(instance_type, validation="prefix") if instance_type else value  # type: ignore[arg-type]
                     )
                 else:
                     # we should not have multiple rdf:type values
                     continue
         if property_values:
             return (
-                identifier,
+                instance_id,
                 property_values,
             )
         else:
@@ -358,6 +382,80 @@ class Queries:
     def count_of_type(self, class_uri: URIRef, named_graph: URIRef | None = None) -> int:
         query = f"SELECT (COUNT(?instance) AS ?instanceCount) WHERE {{ ?instance a <{class_uri}> }}"
         return int(next(iter(self.graph(named_graph).query(query)))[0])  # type: ignore[arg-type, index]
+
+    def types_with_instance_and_property_count(
+        self, remove_namespace: bool = True, named_graph: URIRef | None = None
+    ) -> list[dict[str, Any]]:
+        query = """
+        SELECT ?type (COUNT(DISTINCT ?instance) AS ?instanceCount) (COUNT(DISTINCT ?property) AS ?propertyCount)
+                WHERE {
+                  ?instance a ?type .
+                  ?instance ?property ?value .
+                  FILTER(?property != rdf:type)
+                }
+                GROUP BY ?type
+                ORDER BY DESC(?instanceCount)"""
+        return [
+            {
+                "type": urllib.parse.unquote(remove_namespace_from_uri(type_)) if remove_namespace else type_,
+                "instanceCount": cast(RdfLiteral, instance_count).toPython(),
+                "propertyCount": cast(RdfLiteral, property_count).toPython(),
+            }
+            for type_, instance_count, property_count in list(
+                cast(list[ResultRow], self.graph(named_graph).query(query))
+            )
+        ]
+
+    def properties_with_count(
+        self, remove_namespace: bool = True, named_graph: URIRef | None = None
+    ) -> list[dict[str, Any]]:
+        instance_count_by_type = {
+            entry["type"]: entry["instanceCount"]
+            for entry in self.types_with_instance_and_property_count(remove_namespace=False, named_graph=named_graph)
+        }
+        query = """SELECT ?type ?property (COUNT(DISTINCT ?instance) AS ?instanceCount)
+WHERE {
+  ?instance a ?type .
+  ?instance ?property ?value .
+  FILTER(?property != rdf:type)
+}
+GROUP BY ?type ?property
+ORDER BY ASC(?type) ASC(?property)"""
+        return [
+            {
+                "type": urllib.parse.unquote(remove_namespace_from_uri(type_)) if remove_namespace else type_,
+                "property": urllib.parse.unquote(remove_namespace_from_uri(property)) if remove_namespace else property,
+                "instanceCount": cast(RdfLiteral, instance_count).toPython(),
+                "total": instance_count_by_type[type_],
+            }
+            for type_, property, instance_count in list(cast(list[ResultRow], self.graph(named_graph).query(query)))
+        ]
+
+    @overload
+    def instances_with_properties(
+        self, type: URIRef, remove_namespace: Literal[False], named_graph: URIRef | None = None
+    ) -> dict[URIRef, set[URIRef]]: ...
+
+    @overload
+    def instances_with_properties(
+        self, type: URIRef, remove_namespace: Literal[True], named_graph: URIRef | None = None
+    ) -> dict[str, set[str]]: ...
+
+    def instances_with_properties(
+        self, type: URIRef, remove_namespace: bool = True, named_graph: URIRef | None = None
+    ) -> dict[str, set[str]] | dict[URIRef, set[URIRef]]:
+        query = """SELECT DISTINCT ?instance ?property
+WHERE {{
+    ?instance a <{type}> .
+    ?instance ?property ?value .
+    FILTER(?property != rdf:type)
+}}"""
+        result = defaultdict(set)
+        for instance, property_ in cast(Iterable[ResultRow], self.graph(named_graph).query(query.format(type=type))):
+            instance_str = urllib.parse.unquote(remove_namespace_from_uri(instance)) if remove_namespace else instance
+            property_str = urllib.parse.unquote(remove_namespace_from_uri(property_)) if remove_namespace else property_
+            result[instance_str].add(property_str)
+        return result
 
     def list_instances_ids_by_space(
         self, space_property: URIRef, named_graph: URIRef | None = None